Rename contact search results + add catch to search + add constants folder
