+++ src/stateful_components/Drive/Drive.js
@@ -23,7 +23,6 @@ import {
     copyItems,
     pasteItems,
     renameItem,
-    searchContact,
 } from '../../actions/UserActions';
 import { ClassicSpinner } from 'react-spinners-kit';
 import ToolbarButtons from '../../functional_components/ToolbarButtons';
@@ -211,13 +210,7 @@ class Drive extends React.Component {
     }
 
     componentDidMount() {
-        const {
-            currentItems,
-            currentPath,
-            loadCurrentItems,
-            contactsFound,
-            searchContact,
-        } = this.props;
+        const { currentItems, currentPath, loadCurrentItems } = this.props;
 
         if (!currentItems && !currentPath && !loadCurrentItems) {
             console.log('fetching files');
@@ -227,12 +220,6 @@ class Drive extends React.Component {
                 currentItems
             );
             fetchCurrentItems(currentPath);
-        }
-
-        if (!contactsFound) {
-            searchContact('ludwig');
-        } else {
-            console.log(contactsFound);
         }
     }
 
@@ -640,7 +627,6 @@ const mapStateToProps = (state) => {
         loadDeletion: state.app.loadDeletion,
         clipboard: state.app.clipboard,
         loadPaste: state.app.loadPaste,
-        contactsFound: state.app.contactsFound,
     };
 };
 
@@ -656,7 +642,6 @@ export default withRouter(
             copyItems,
             pasteItems,
             renameItem,
-            searchContact,
         }
     )(Drive)
 );
